Improved some stuff
I removed the nice stats info for something else because system("cls") is apparently "resource heavy, insecure and can be recognized as a false positive by anti virus programs." so to remove those negative points i will just make a simple alternative that doesn't utilize the system function. This also makes stuff more portable

I Also changed some regex and made a #define for the working proxies file name so if anyone wants to change it then it can be easily done in the source.